requirements-ai.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
google-genai==1.5.0
